requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1

deep_sort_realtime==1.3.2
deepface==0.0.93
mediapipe==0.10.15
numpy==1.24.3
pinecone==5.3.1
pymongo==4.10.1
pyserial==3.5
python-dotenv==1.0.1